app/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
httpx==0.27.2
pydantic==2.9.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
